feat(linux): ship real focus tracking in CLI + X11 window-text capture
Complete Linux focus tracking so released CLIs witness the active window
instead of falling through to the degraded stub monitor.

- Forward x11/wayland/dbus features from cpoe_cli to the cpoe engine via a
  Linux-target dependency table, so every Linux build -- including the
  deb/AppImage/rpm packaging builds that just run `cargo build` -- ships the
  real X11/Wayland/DBus active-window providers by default.
- Implement the X11 path for WindowTextCapture (window_text.rs): enumerate
  _NET_CLIENT_LIST and read _NET_WM_NAME/WM_NAME + _NET_WM_PID. X11 exposes
  no API to read other windows' text content, so the text-content readers
  return None and text_content stays empty; a fallback stub covers
  Linux-without-x11 and other unixes.
- Make `cargo clippy --workspace -- -D warnings` clean under the focus
  features. The Linux-only code (plus some always-compiled style nits) was
  never linted by the macOS CI; auto-fixed mechanical lints and added Debug
  derives / targeted allows. No behavior change.

+++ crates/cpoe/src/ffi/evidence_export.rs
@@ -228,7 +228,7 @@ pub(crate) fn build_wire_packet_with_ai(
                     input: vdf_input_bytes,
                     merkle_root,
                     sampled_proofs: vec![],
-                    claimed_duration: ev.vdf_iterations.saturating_mul(1000) / ips as u64,
+                    claimed_duration: ev.vdf_iterations.saturating_mul(1000) / ips,
                 },
                 jitter_binding: None,
                 physical_state: None,
